refactor zipkin code to an issolated module in core
