+++ src/core/index.ts
@@ -1,5 +1,7 @@
 export * from "./schema-stitching";
 export * from "./config";
 export * from "./server";
-export * from "./plugins";
-export { zipkinFetcher } from "./zipkin";
+export * from "./plugins/graphql";
+export * from "./zipkin/apollo-link-zipkin";
+export * from "./zipkin/plugin";
+export * from "./zipkin/zipkin-instrumentation-hapi";
