--- src/core/plugins/index.ts
@@ -1,2 +1,0 @@
-export * from "./graphql";
-export * from "./zipkin";
